Add new test cases that expose bug in reference solution.
To reproduce, run these commands (using the reference algorithm):

./bin/sgdf headless -s sgdf/benchmarking/examples/bar-source.png -t sgdf/benchmarking/examples/bar-target.png -m sgdf/benchmarking/examples/bar-mask.png -o output.png
./bin/sgdf headless -s sgdf/benchmarking/examples/bar-source.png -t sgdf/benchmarking/examples/bar-target.png -m sgdf/benchmarking/examples/bar-mask-2.png -o output-2.png

The source image contains a very slight change in brightness.

The target image is all black.

When we fuse the 2 images, we expect a slight change in brightness in the result. However, the buggy
algorithm will create a large change in brightness.

This is because the sparse matrix used in the reference algorithm contains a stray constraint, in
the case where a pixel in the masked area is on the edge (top, left, right, or bottom) of the source
image. The stray constraint is of the form v[N] = s[N] -- where v is the result and s is the source.

We already have constraints of the form v[N] = t[N] on the border of the mask. This is ok. However,
source image pixels should never be compared directly with the resulting pixels. We only care about
the gradient of the source image.

This effect is particularly noticeable because in this example, the source image is much brighter
than the target image.

See visual demonstration:
https://www.dropbox.com/sh/66fp75xwilfhn0v/AACZlg0Ce8O9POlJhhLzabgFa?lst